Fixed console outputs and response

+++ src/modules/loadFromRemote.ts
@@ -4,7 +4,9 @@ import fs from "fs";
 import path from "path";
 import https from "https";
 import Listr from "listr";
+import { safeDirectoryCreate, safeDirectoryRemove } from "../utils/fileUtils";
 
+const ora = require("ora");
 const root = "scripts/";
 const rootS3 = "https://cloudfront.collapp.live/plugins/";
 
@@ -29,12 +31,11 @@ const downloadFileToLocalDirectory = async (plugin: string) => {
   const p = path.join(__dirname, root, plugin);
 
   // If local version exists, delete it
-  if (fs.existsSync(p)) {
-    fs.rmdirSync(p, { recursive: true });
-  }
+  safeDirectoryRemove(p);
 
   // Create new local version of plugin
-  fs.mkdirSync(p);
+  safeDirectoryCreate(p);
+
   const file = fs.createWriteStream(p + "/server.js");
   https
     .get(rootS3 + plugin + "/server.js", (response) => {
@@ -45,21 +46,19 @@ const downloadFileToLocalDirectory = async (plugin: string) => {
     })
     .on("error", (err) => {
       fs.unlink(p, () => {});
-      fs.rmdirSync(p, { recursive: true });
+      safeDirectoryRemove(p);
       console.log(chalk.red(err));
     });
 };
 
 const deleteUnwanted = async (plugin: string) => {
-  fs.rmdirSync(path.resolve(__dirname, root, plugin), { recursive: true });
+  safeDirectoryRemove(path.resolve(__dirname, root, plugin));
 };
 
 export async function syncPlugins() {
-  console.log("Let me check if we are up to date...");
+  const spinner = ora("Loading remote modules\n").start();
 
-  if (!fs.existsSync(path.resolve(__dirname, root))) {
-    fs.mkdirSync(path.resolve(__dirname, root));
-  }
+  safeDirectoryCreate(path.resolve(__dirname, root));
 
   const params = {
     Bucket: process.env.AWS_BUCKET,
@@ -79,13 +78,16 @@ export async function syncPlugins() {
   let localPlugins = fs.readdirSync(path.resolve(__dirname, "./scripts/"));
 
   const toDelete = setDifference(localPlugins, remotePlugins);
-  const tasksDelete = new Listr(
-    toDelete.map((d) => ({
-      title: chalk.red.bold("Removed") + ` local version of '${d}' plugin`,
-      task: () => deleteUnwanted(d),
-    }))
-  );
-  await tasksDelete.run();
+
+  if (toDelete.length > 0) {
+    const tasksDelete = new Listr(
+      toDelete.map((d) => ({
+        title: chalk.red.bold("Removed") + ` local version of '${d}' plugin`,
+        task: () => deleteUnwanted(d),
+      }))
+    );
+    await tasksDelete.run();
+  }
 
   localPlugins = localPlugins.filter((local) => !toDelete.includes(local));
 
@@ -98,14 +100,16 @@ export async function syncPlugins() {
     ...emptyPlugins,
   ];
 
-  const tasksDownload = new Listr(
-    toDownload.map((d) => ({
-      title: chalk.green.bold("Downloaded") + ` '${d}' plugin`,
-      task: () => downloadFileToLocalDirectory(d),
-    }))
-  );
-  await tasksDownload.run();
-
-  if (toDownload.length == 0 && toDelete.length == 0)
-    console.log(chalk.blue.bold("Uff, all seems to be up to date"));
+  if (toDownload.length == 0 && toDelete.length == 0) {
+    spinner.succeed("All is already up to date");
+  } else {
+    const tasksDownload = new Listr(
+      toDownload.map((d) => ({
+        title: chalk.green.bold("Download ->") + ` '${d}' plugin`,
+        task: () => downloadFileToLocalDirectory(d),
+      }))
+    );
+    await tasksDownload.run();
+    spinner.succeed("Now it's all up to date");
+  }
 }
